# docs/openapi.yaml
openapi: 3.1.0
info:
  title: P2P framework
  version: 0.2
paths:
...
  /append/:key:
    post:
      summary: Append to an entry in the hash table
      description: data shall be an array (if not it will be encapsulated in one) and will be concatenated with the existing entry if there is one
      parameters:
      - name: key
        in: path
      - name: data
        in: body
        schema:
          format:
            data: "any"
      responses:
        "200":
          description: OK
          content:
            schema:
              format:
                error: "boolean"
                key: "string"
                value: "any[]"
              example:
                error: false
                key: "Grundgesetz"
                value: "[ 'inserted', 'data' ]"
        "400":
          description: User error
          content:
            schema:
              format:
                error: "boolean"
                errorMsg: "string"
              example:
                error: true
                errorMsg: "missing data"
...